requirements.txt
amqp==2.2.1
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.5
django-celery-beat==1.0.1
django-cors-headers==2.1.0
django-environ==0.4.4
django-filter==1.0.4
django-nose==1.4.5
djangorestframework==3.6.4
idna==2.6
kombu==4.1.0
model-mommy==1.4.0
nose==1.3.7
psycopg2==2.7.3.1
pyfcm==1.4.2
pytz==2017.2
requests==2.18.4
requests-toolbelt==0.8.0
six==1.10.0
urllib3==1.22
vine==1.1.4
